+++ test/unit/version_test.lua
@@ -1,0 +1,57 @@
+local fio = require('fio')
+
+local t = require('luatest')
+local g = t.group('version')
+
+local helpers = require('test.helper')
+
+g.before_all = function()
+    g.cluster = helpers.Cluster:new({
+        datadir = fio.tempdir(),
+        server_command = helpers.entrypoint('srv_say_hi'),
+        use_vshard = true,
+        replicasets = {
+            {
+                uuid = helpers.uuid('a'),
+                alias = 'router',
+                roles = { 'crud-router' },
+                servers = {
+                    { instance_uuid = helpers.uuid('a', 1), alias = 'router' },
+                },
+            },
+            {
+                uuid = helpers.uuid('b'),
+                alias = 's-1',
+                roles = { 'crud-storage' },
+                servers = {
+                    { instance_uuid = helpers.uuid('b', 1), alias = 's1-master' },
+                    { instance_uuid = helpers.uuid('b', 2), alias = 's1-replica' },
+                },
+            },
+            {
+                uuid = helpers.uuid('c'),
+                alias = 's-2',
+                roles = { 'crud-storage' },
+                servers = {
+                    { instance_uuid = helpers.uuid('c', 1), alias = 's2-master' },
+                    { instance_uuid = helpers.uuid('c', 2), alias = 's2-replica' },
+                },
+            }
+        },
+    })
+    g.cluster:start()
+end
+
+g.after_all = function()
+    g.cluster:stop()
+    fio.rmtree(g.cluster.datadir)
+end
+
+g.test_version = function()
+    local handle = io.popen('git describe --tags --always')
+    local version = handle:read("*a"):gsub('\n*', '')
+    handle:close()
+    t.assert_equals(g.cluster.main_server.net_box:eval('return crud.VERSION'), version)
+    t.assert_equals(g.cluster.servers[2].net_box:eval('return _crud.VERSION'), version)
+    t.assert_equals(g.cluster.servers[3].net_box:eval('return _crud.VERSION'), version)
+end